fix bug & optimize

+++ backend/super-bi/src/main/java/net/aopacloud/superbi/model/dto/DashboardComponentDTO.java
@@ -16,7 +16,7 @@ public class DashboardComponentDTO {
 
     private String type;
 
-    private Integer reportId;
+    private Long reportId;
 
     private Date createTime;
